+++ ReactiveCocoaTests/UIKit/UISearchBarSpec.swift
@@ -9,12 +9,18 @@ class UISearchBarSpec: QuickSpec {
 	override func spec() {
 		var searchBar: UISearchBar!
 		weak var _searchBar: UISearchBar?
-
+		var receiver: SearchBarDelegateReceiver!
 
 		beforeEach {
 			autoreleasepool {
+				receiver = SearchBarDelegateReceiver()
 				searchBar = UISearchBar(frame: .zero)
 				_searchBar = searchBar
+
+				_ = searchBar.reactive.textValues
+				searchBar.delegate = receiver
+
+				expect(searchBar.delegate).toNot(beIdenticalTo(receiver))
 			}
 		}
 
@@ -49,9 +55,13 @@ class UISearchBarSpec: QuickSpec {
 				latestValue = text
 			}
 
+			expect(latestValue).to(beNil())
+			expect(receiver.endEditingTexts.isEmpty) == true
+
 			searchBar.delegate!.searchBarTextDidEndEditing!(searchBar)
 
 			expect(latestValue) == searchBar.text
+			expect(receiver.endEditingTexts.last) == searchBar.text
 		}
 
 		it("should emit user initiated changes to its text value continuously") {
@@ -62,83 +72,142 @@ class UISearchBarSpec: QuickSpec {
 				latestValue = text
 			}
 
+			expect(latestValue).to(beNil())
+			expect(receiver.texts.isEmpty) == true
+
 			searchBar.delegate!.searchBar!(searchBar, textDidChange: "newValue")
 			expect(latestValue) == "newValue"
+			expect(receiver.texts.last) == "newValue"
 		}
 
-		it("should pass through the intercepted calls") {
-			searchBar.text = "newValue"
+		it("should accept changes from bindings to its scope button index") {
+			let firstChange = 1
+			let secondChange = 2
 
-			var latestValue: String?
-			searchBar.reactive.continuousTextValues.observeValues { text in
+			searchBar.scopeButtonTitles = ["First", "Second", "Third"]
+
+			let (pipeSignal, observer) = Signal<Int, NoError>.pipe()
+			searchBar.reactive.selectedScopeButtonIndex <~ SignalProducer(pipeSignal)
+
+			observer.send(value: firstChange)
+			expect(searchBar.selectedScopeButtonIndex) == firstChange
+
+			observer.send(value: secondChange)
+			expect(searchBar.selectedScopeButtonIndex) == secondChange
+		}
+
+		it("should emit user initiated changes to its text value when the editing ends") {
+			searchBar.scopeButtonTitles = ["First", "Second", "Third"]
+
+			var latestValue: Int?
+			searchBar.reactive.selectedScopeButtonIndices.observeValues { text in
 				latestValue = text
 			}
 
-			let receiver = SearchBarDelegateReceiver()
-			searchBar.delegate = receiver
-			expect(receiver.textDidChangeCounter) == 0
+			expect(latestValue).to(beNil())
+			expect(receiver.selectedScopeButtonIndices.isEmpty) == true
 
-			searchBar.delegate!.searchBar!(searchBar, textDidChange: "newValue")
-			expect(latestValue) == "newValue"
-			expect(receiver.textDidChangeCounter) == 1
+			searchBar.delegate!.searchBar!(searchBar, selectedScopeButtonIndexDidChange: 1)
+
+			expect(latestValue) == 1
+			expect(receiver.selectedScopeButtonIndices.last) == 1
+
+			searchBar.delegate!.searchBar!(searchBar, selectedScopeButtonIndexDidChange: 2)
+
+			expect(latestValue) == 2
+			expect(receiver.selectedScopeButtonIndices.last) == 2
 		}
 
-		it("should pass through the unintercepted calls") {
-			searchBar.reactive.continuousTextValues.observe { _ in }
+		it("should notify when the cancel button is clicked") {
+			var isClicked: Bool?
+			searchBar.reactive.cancelButtonClicked
+				.observeValues { isClicked = true }
 
-			let receiver = SearchBarDelegateReceiver()
-			searchBar.delegate = receiver
-			expect(receiver.searchButtonClickedCounter) == 0
+			expect(isClicked).to(beNil())
+			expect(receiver.cancelButtonClickedCounter) == 0
 
-			searchBar.delegate!.searchBarSearchButtonClicked!(searchBar)
-			expect(receiver.searchButtonClickedCounter) == 1
-		}
-
-		it("should preserve the original delegate, and pass through the unintercepted calls") {
-			let receiver = SearchBarDelegateReceiver()
-			searchBar.delegate = receiver
-			expect(receiver.searchButtonClickedCounter) == 0
-
-			searchBar.reactive.continuousTextValues.observe { _ in }
-			expect(receiver.searchButtonClickedCounter) == 0
-
-			searchBar.delegate!.searchBarSearchButtonClicked!(searchBar)
-			expect(receiver.searchButtonClickedCounter) == 1
-		}
-		
-		it("should pass through the unintercepted calls") {
-			searchBar.reactive.continuousTextValues.observe { _ in }
-			
-			let receiver = SearchBarDelegateReceiver()
-			searchBar.delegate = receiver
-			expect(receiver.searchBarCancelButtonClickedCounter) == 0
-			
 			searchBar.delegate!.searchBarCancelButtonClicked!(searchBar)
-			expect(receiver.searchBarCancelButtonClickedCounter) == 1
+			expect(isClicked) == true
+			expect(receiver.cancelButtonClickedCounter) == 1
 		}
-		
+
+		it("should notify when the search button is clicked") {
+			var isClicked: Bool?
+			searchBar.reactive.searchButtonClicked
+				.observeValues { isClicked = true }
+
+			expect(isClicked).to(beNil())
+			expect(receiver.searchButtonClickedCounter) == 0
+
+			searchBar.delegate!.searchBarSearchButtonClicked!(searchBar)
+			expect(isClicked) == true
+			expect(receiver.searchButtonClickedCounter) == 1
+		}
+
+
+		it("should notify when the bookmark button is clicked") {
+			var isClicked: Bool?
+			searchBar.reactive.bookmarkButtonClicked
+				.observeValues { isClicked = true }
+
+			expect(isClicked).to(beNil())
+			expect(receiver.bookmarkButtonClickedCounter) == 0
+
+			searchBar.delegate!.searchBarBookmarkButtonClicked!(searchBar)
+			expect(isClicked) == true
+			expect(receiver.bookmarkButtonClickedCounter) == 1
+		}
+
+
+		it("should notify when the results list button is clicked") {
+			var isClicked: Bool?
+			searchBar.reactive.resultsListButtonClicked
+				.observeValues { isClicked = true }
+
+			expect(isClicked).to(beNil())
+			expect(receiver.resultsListButtonClickedCounter) == 0
+
+			searchBar.delegate!.searchBarResultsListButtonClicked!(searchBar)
+			expect(isClicked) == true
+			expect(receiver.resultsListButtonClickedCounter) == 1
+		}
 	}
 }
 
 class SearchBarDelegateReceiver: NSObject, UISearchBarDelegate {
-	var textDidChangeCounter = 0
-	var textDidEndEditingCounter = 0
+	var texts: [String] = []
+	var endEditingTexts: [String] = []
 	var searchButtonClickedCounter = 0
-	var searchBarCancelButtonClickedCounter = 0
+	var cancelButtonClickedCounter = 0
+	var bookmarkButtonClickedCounter = 0
+	var resultsListButtonClickedCounter = 0
+	var selectedScopeButtonIndices: [Int] = []
 
 	func searchBarSearchButtonClicked(_ searchBar: UISearchBar) {
 		searchButtonClickedCounter += 1
 	}
 
 	func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
-		textDidChangeCounter += 1
+		texts.append(searchText)
 	}
 
 	func searchBarTextDidEndEditing(_ searchBar: UISearchBar) {
-		textDidEndEditingCounter += 1
+		endEditingTexts.append(searchBar.text ?? "")
 	}
 	
 	func searchBarCancelButtonClicked(_ searchBar: UISearchBar) {
-		searchBarCancelButtonClickedCounter += 1
+		cancelButtonClickedCounter += 1
+	}
+
+	func searchBarResultsListButtonClicked(_ searchBar: UISearchBar) {
+		resultsListButtonClickedCounter += 1
+	}
+
+	func searchBarBookmarkButtonClicked(_ searchBar: UISearchBar) {
+		bookmarkButtonClickedCounter += 1
+	}
+
+	func searchBar(_ searchBar: UISearchBar, selectedScopeButtonIndexDidChange selectedScope: Int) {
+		selectedScopeButtonIndices.append(selectedScope)
 	}
 }